tools/pyproject.toml
[tool.poetry]
name = "tools"
version = "0.1.0"
description = ""
authors = ["Damodar Dahal <damo94761@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dateutil = "^2.9.0.post0"
nepali = "^1.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
